airflow/requirements.txt
azure-ai-ml
azure-core
airflow-provider-azure-machinelearning
